:sparkles: Adicionada validacao para nao delecao de livros indisponiveis e teste da validacao

+++ src/modules/books/useCases/deleteBook/DeleteBookUseCase.ts
@@ -1,4 +1,5 @@
 import { IBookRepository } from "@modules/books/repositories/IBookRepository";
+import { AppError } from "@shared/errors/AppError";
 import { inject, injectable } from "tsyringe";
 
 @injectable()
@@ -10,6 +11,12 @@ class DeleteBookUseCase {
     ){}
 
     async execute(id: string) {
+        const book = await this.booksRepository.findById(id);
+
+        if (!book?.available) {
+            throw new AppError("unavailable book!")
+        }
+
         return await this.booksRepository.delete(id)
     }
